:bug: restrict all_questions() to Quizzes
The responseblock code was clearly written when the only blocks in Match with `needs_submit()`
where Quizblocks. Since then, the `CounselingReferral` has crept in. That causes
`all_questions()` to fail as seen here:

https://sentry.ccnmtl.columbia.edu/sentry-internal/match/group/696/

breaking edit pages with a responseblock.

Long term, responseblock should probably use some more abstract approach to getting
the possible questions to choose from so that it could work with any block type
that provides the right API.

In the meantime, just pin it down and make it exclusive to Quizblock's Quizzes.